feat(pagination): add basic test
